Moved types used by tests to CommonTestTypes, unifing them
