+++ swagger2.cabal
@@ -105,6 +105,7 @@ test-suite spec
   other-modules:
     SpecCommon
     Data.SwaggerSpec
+    Data.Swagger.CommonTestTypes
     Data.Swagger.ParamSchemaSpec
     Data.Swagger.SchemaSpec
     Data.Swagger.Schema.ValidationSpec
