+++ test/Data/Swagger/ParamSchemaSpec.hs
@@ -14,6 +14,7 @@ import GHC.Generics
 import Data.Swagger
 import Data.Swagger.Internal (SwaggerKind(..))
 
+import Data.Swagger.CommonTestTypes
 import SpecCommon
 import Test.Hspec
 
@@ -32,85 +33,3 @@ spec = do
 
 main :: IO ()
 main = hspec spec
-
--- ========================================================================
--- Unit type
--- ========================================================================
-
-data Unit = Unit deriving (Generic)
-instance ToParamSchema Unit
-
-unitSchemaJSON :: Value
-unitSchemaJSON = [aesonQQ|
-{
-  "type": "string",
-  "enum": ["Unit"]
-}
-|]
-
--- ========================================================================
--- Color (enum)
--- ========================================================================
-data Color
-  = Red
-  | Green
-  | Blue
-  deriving (Generic)
-instance ToParamSchema Color
-
-colorSchemaJSON :: Value
-colorSchemaJSON = [aesonQQ|
-{
-  "type": "string",
-  "enum": ["Red", "Green", "Blue"]
-}
-|]
-
--- ========================================================================
--- Status (constructorTagModifier)
--- ========================================================================
-
-data Status = StatusOk | StatusError deriving (Generic)
-
-instance ToParamSchema Status where
-  toParamSchema = genericToParamSchema defaultSchemaOptions
-    { constructorTagModifier = map toLower . drop (length "Status") }
-
-statusSchemaJSON :: Value
-statusSchemaJSON = [aesonQQ|
-{
-  "type": "string",
-  "enum": ["ok", "error"]
-}
-|]
-
--- ========================================================================
--- Email (newtype with unwrapUnaryRecords set to True)
--- ========================================================================
-
-newtype Email = Email { getEmail :: String }
-  deriving (Generic)
-instance ToParamSchema Email
-
-emailSchemaJSON :: Value
-emailSchemaJSON = [aesonQQ|
-{
-  "type": "string"
-}
-|]
-
--- ========================================================================
--- UserId (non-record newtype)
--- ========================================================================
-
-newtype UserId = UserId Integer
-  deriving (Generic)
-instance ToParamSchema UserId
-
-userIdSchemaJSON :: Value
-userIdSchemaJSON = [aesonQQ|
-{
-  "type": "integer"
-}
-|]
-
